catching up with Wt

+++ src/eu/webtoolkit/jwt/chart/WAxis.java
@@ -693,7 +693,7 @@ public class WAxis {
 	}
 
 	int getSegmentCount() {
-		return this.segments_.size();
+		return (int) this.segments_.size();
 	}
 
 	private WCartesianChart chart_;
